Make SignalWrapper an enum and represent InstanceStats there. Move run-time panic check to compile-time type power.
