+++ crates/conductor_lib/src/signal_wrapper.rs
@@ -1,10 +1,17 @@
-use holochain_core::signal::Signal;
+use holochain_core::{context::InstanceStats, signal::Signal};
 use holochain_json_api::{error::JsonError, json::JsonString};
+use std::collections::HashMap;
 
-/// This struct wraps a Signal from core before serializing and sending over
+/// This enum wraps a Signal from core before serializing and sending over
 /// an interface to the UI or other client.
 #[derive(Serialize, Deserialize, Debug, DefaultJson)]
-pub struct SignalWrapper {
-    pub signal: Signal,
-    pub instance_id: String,
+#[allow(clippy::large_enum_variant)]
+pub enum SignalWrapper {
+    InstanceSignal {
+        signal: Signal,
+        instance_id: String,
+    },
+    InstanceStats {
+        instance_stats: HashMap<String, InstanceStats>,
+    },
 }
